Lex strings
This _looks_ good, but if we want to put in a literal " it needs to be
quoted.

There are a couple of problems here, the first is that we iterate one
by one, unless we hit a symbol in which case we jump.

One nice feature I'd like is the ability to do a triple quote to evade
this issue.