Add ranting, arithmetic symbol recognition
regression test for single digits, pushed parse_token out to its own
function

Extending this has been pretty easy so far. Granted it's a simple
language and I'm parsing a tiny  subset of it.

One issue I see with adding arithmetic is that it adds a bunch of
symbol freedom. That's... kind of okay, but it definitely is worth
noting that symbols+like/this are valid lisp symbols now.