Make number priority explicit, turn parse_token into a function
Not a generator.  Interesting problem here is that if there's ever a
mismatch between the lexer and the parser, (or heavens forbid, a bug)
then this just returns none, which will be in the parse output stream.

No bueno.

Two solutions would be to:
1) Bite the bullet and write a parser combinator
2) Have no bugs.

This has some relation to what's going on at work, and making sure
that code follows certain standards. One possibility is to attempt to
verify that the code itself works well, but requires manual
intervention.

The imho correct solution would take weeks of figuring out how code
generation works, but this would remove the tedium of writing code
that may be incorrect and instead write code that is 100% right,
provided the code generation is correct.