Java question resolved
Java is pass by reference by value. This means the reference is copied
in, and if the variable is final (or effectively final) then the
reference's value is immutable throughout the lifetime of the
reference.

The object referenced by the... reference can be as mutable as it
wants to be, since all running code refers to the same value they all
share the same meaning.