Excise string splicing
This makes the parameter signature a little less pretty, but to
mitigate that would be to pull that out into a namedtuple which I have
littered all over this codebase.

Otherwise, major win. Now parsers just share the reference to an
immutable string and they themselves move the index instead of
consuming tokens one at a time from the head of a potentially giant string.