Attempt to make progress
One of the more aimless commits. make_parser is a HUGE mess, but I
wanted to commit this before I nuke it into something significantly
easier to reason about.

In particular I keep defining data structures that complicate
implementations. My combinators then suffer from increased
complexity. They shouldn't care about the ds of an individual parser,
but 'and' specifically is a problem because it has to know how far the
parser has moved.

Hm. Is the recursion an issue? I don't think so, I think it's easier
to apply recursive rules instead of... I've written this before I'm sure.