Rubber duck forever
Major decision to make is should this stage handle ast structure?

The answer is probably not, but at this point if we wanted to go for
gold and try for a metacompiler this is the point where we'd be
generating code. So it might be valuable to consolidate this so that
we could easily translate all these sub calls into emitting code
instead of doing runtime compiler building.