Deliberate effects of memoize
If we've got a character stream that we iterate over, we might not
care about the index. In any case, we don't want to have to deal with
moving through the string ourselves, if we do care about the index of
the string we can always iterate over the enumerate wrapper which'll
give us the index but will handle the index problem.

I wonder what the use case of a pushback data stream is (Java thing)