Dispatch attempt.
Now we have some delay semantics for looking up identifiers.

One of the issues with this is that the parser just globs up
identifiers as if they're normal parsers, where they're not.

Would really like to start moving to a typed language right about
now. Kotlin is a strong candidate.